Weapon & Armour balanced reward system
Added code that rewards either weapons or armour from any loaded mod.

To make it balanced it uses the players level and compares the stats of the item to the vanilla armour/damage scaling.

It scales from level 1-40, so for example at level one you will get items similar in power to Iron and at 40 you will get items similar in power to daedric.

The only weapons that currently aren't used are staffs and Daedric Artifacts. These are removed via keyword checks.